Migrate `src/api/translations` to TypeScript.
